requirements.txt
python_gitlab==2.0.1